move tx functions out of testing functions
