+++ tests/e2e/e2e_cw_test.go
@@ -5,13 +5,10 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"gopkg.in/yaml.v2"
 	"os"
 	"path/filepath"
 	"strconv"
 	"time"
-
-	"github.com/cosmos/cosmos-sdk/client/flags"
 )
 
 func (s *IntegrationTestSuite) testCWCounter() {
@@ -52,135 +49,4 @@ func (s *IntegrationTestSuite) testCWCounter() {
 	err = json.Unmarshal(data, &counterResp)
 	s.Require().NoError(err)
 	s.Require().Equal(1, counterResp["count"])
-}
-
-func (s *IntegrationTestSuite) storeWasm(ctx context.Context, c *chain, valIdx int, sender, wasmPath string) string {
-	storeCmd := []string{
-		gaiadBinary,
-		txCommand,
-		"wasm",
-		"store",
-		wasmPath,
-		fmt.Sprintf("--from=%s", sender),
-		fmt.Sprintf("--%s=%s", flags.FlagFees, standardFees.String()),
-		fmt.Sprintf("--%s=%s", flags.FlagChainID, c.id),
-		"--gas=5000000",
-		"--keyring-backend=test",
-		"--broadcast-mode=sync",
-		"--output=json",
-		"-y",
-	}
-
-	s.T().Logf("%s storing wasm on host chain %s", sender, s.chainB.id)
-	s.executeGaiaTxCommand(ctx, c, storeCmd, valIdx, s.defaultExecValidation(c, valIdx))
-	s.T().Log("successfully sent store wasm tx")
-	contractsCounter++
-	return strconv.Itoa(contractsCounter)
-}
-
-func (s *IntegrationTestSuite) instantiateWasm(ctx context.Context, c *chain, valIdx int, sender, codeID,
-	msg, label string,
-) string {
-	storeCmd := []string{
-		gaiadBinary,
-		txCommand,
-		"wasm",
-		"instantiate",
-		codeID,
-		msg,
-		fmt.Sprintf("--from=%s", sender),
-		fmt.Sprintf("--%s=%s", flags.FlagFees, standardFees.String()),
-		fmt.Sprintf("--%s=%s", flags.FlagChainID, c.id),
-		fmt.Sprintf("--label=%s", label),
-		"--no-admin",
-		"--gas=500000",
-		"--keyring-backend=test",
-		"--broadcast-mode=sync",
-		"--output=json",
-		"-y",
-	}
-
-	s.T().Logf("%s instantiating wasm on host chain %s", sender, s.chainB.id)
-	s.executeGaiaTxCommand(ctx, c, storeCmd, valIdx, s.defaultExecValidation(c, valIdx))
-	s.T().Log("successfully sent instantiate wasm tx")
-	chainEndpoint := fmt.Sprintf("http://%s", s.valResources[c.id][0].GetHostPort("1317/tcp"))
-	address, err := queryWasmContractAddress(chainEndpoint, sender, contractsCounterPerSender[sender])
-	s.Require().NoError(err)
-	contractsCounterPerSender[sender]++
-	return address
-}
-
-func (s *IntegrationTestSuite) instantiate2Wasm(ctx context.Context, c *chain, valIdx int, sender, codeID,
-	msg, salt, label string,
-) string {
-	storeCmd := []string{
-		gaiadBinary,
-		txCommand,
-		"wasm",
-		"instantiate2",
-		codeID,
-		msg,
-		salt,
-		fmt.Sprintf("--from=%s", sender),
-		fmt.Sprintf("--%s=%s", flags.FlagFees, standardFees.String()),
-		fmt.Sprintf("--%s=%s", flags.FlagChainID, c.id),
-		fmt.Sprintf("--label=%s", label),
-		"--no-admin",
-		"--gas=250000",
-		"--keyring-backend=test",
-		"--broadcast-mode=sync",
-		"--output=json",
-		"-y",
-	}
-
-	s.T().Logf("%s instantiating wasm on host chain %s", sender, s.chainB.id)
-
-	s.executeGaiaTxCommand(ctx, c, storeCmd, valIdx, s.defaultExecValidation(c, valIdx))
-	s.T().Log("successfully sent instantiate2 wasm tx")
-	chainEndpoint := fmt.Sprintf("http://%s", s.valResources[c.id][0].GetHostPort("1317/tcp"))
-	address, err := queryWasmContractAddress(chainEndpoint, sender, contractsCounterPerSender[sender])
-	s.Require().NoError(err)
-	contractsCounterPerSender[sender]++
-	return address
-}
-
-func (s *IntegrationTestSuite) executeWasm(ctx context.Context, c *chain, valIdx int, sender, addr, msg string) {
-	execCmd := []string{
-		gaiadBinary,
-		txCommand,
-		"wasm",
-		"execute",
-		addr,
-		msg,
-		fmt.Sprintf("--from=%s", sender),
-		fmt.Sprintf("--%s=%s", flags.FlagFees, standardFees.String()),
-		fmt.Sprintf("--%s=%s", flags.FlagChainID, c.id),
-		"--gas=250000",
-		"--keyring-backend=test",
-		"--broadcast-mode=sync",
-		"--output=json",
-		"-y",
-	}
-	s.T().Logf("%s executing wasm on host chain %s", sender, s.chainB.id)
-	s.executeGaiaTxCommand(ctx, c, execCmd, valIdx, s.defaultExecValidation(c, valIdx))
-	s.T().Log("successfully sent execute wasm tx")
-}
-
-func (s *IntegrationTestSuite) queryBuildAddress(ctx context.Context, c *chain, valIdx int, codeHash, creatorAddress, saltHexEncoded string,
-) (res string) {
-	cmd := []string{
-		gaiadBinary,
-		queryCommand,
-		"wasm",
-		"build-address",
-		codeHash,
-		creatorAddress,
-		saltHexEncoded,
-	}
-
-	s.executeGaiaTxCommand(ctx, c, cmd, valIdx, func(stdOut []byte, stdErr []byte) bool {
-		s.Require().NoError(yaml.Unmarshal(stdOut, &res))
-		return true
-	})
-	return res
 }
