+++ tests/e2e/e2e_setup_test.go
@@ -38,90 +38,9 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	authvesting "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-
-	"github.com/cosmos/gaia/v23/types"
 )
-
-const (
-	gaiadBinary    = "gaiad"
-	txCommand      = "tx"
-	queryCommand   = "query"
-	keysCommand    = "keys"
-	gaiaHomePath   = "/home/nonroot/.gaia"
-	photonDenom    = "photon"
-	uatomDenom     = types.UAtomDenom
-	stakeDenom     = "stake"
-	initBalanceStr = "110000000000stake,100000000000000000photon,100000000000000000uatom"
-	minGasPrice    = "0.005"
-	// the test basefee in genesis is the same as minGasPrice
-	// global fee lower/higher than min_gas_price
-	initialBaseFeeAmt               = "0.005"
-	gas                             = 200000
-	govProposalBlockBuffer          = 35
-	relayerAccountIndexHermes       = 0
-	numberOfEvidences               = 10
-	slashingShares            int64 = 10000
-
-	proposalMaxTotalBypassFilename   = "proposal_max_total_bypass.json"
-	proposalCommunitySpendFilename   = "proposal_community_spend.json"
-	proposalLSMParamUpdateFilename   = "proposal_lsm_param_update.json"
-	proposalBlocksPerEpochFilename   = "proposal_blocks_per_epoch.json"
-	proposalFailExpedited            = "proposal_fail_expedited.json"
-	proposalExpeditedSoftwareUpgrade = "proposal_expedited_software_upgrade.json"
-	proposalSoftwareUpgrade          = "proposal_software_upgrade.json"
-	proposalCancelSoftwareUpgrade    = "proposal_cancel_software_upgrade.json"
-
-	// proposalAddConsumerChainFilename    = "proposal_add_consumer.json"
-	// proposalRemoveConsumerChainFilename = "proposal_remove_consumer.json"
-
-	hermesBinary              = "hermes"
-	hermesConfigWithGasPrices = "/root/.hermes/config.toml"
-	hermesConfigNoGasPrices   = "/root/.hermes/config-zero.toml"
-	transferPort              = "transfer"
-	transferChannel           = "channel-0"
-
-	v2TransferClient = "08-wasm-1"
-
-	govAuthority = "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn"
-)
-
-var (
-	gaiaConfigPath            = filepath.Join(gaiaHomePath, "config")
-	stakingAmount             = math.NewInt(100000000000)
-	stakingAmountCoin         = sdk.NewCoin(uatomDenom, stakingAmount)
-	tokenAmount               = sdk.NewCoin(uatomDenom, math.NewInt(3300000000)) // 3,300uatom
-	standardFees              = sdk.NewCoin(uatomDenom, math.NewInt(330000))     // 0.33uatom
-	depositAmount             = sdk.NewCoin(uatomDenom, math.NewInt(330000000))  // 3,300uatom
-	distModuleAddress         = authtypes.NewModuleAddress(distrtypes.ModuleName).String()
-	govModuleAddress          = authtypes.NewModuleAddress(govtypes.ModuleName).String()
-	proposalCounter           = 0
-	contractsCounter          = 0
-	contractsCounterPerSender = map[string]uint64{}
-)
-
-type IntegrationTestSuite struct {
-	suite.Suite
-
-	tmpDirs        []string
-	chainA         *chain
-	chainB         *chain
-	dkrPool        *dockertest.Pool
-	dkrNet         *dockertest.Network
-	hermesResource *dockertest.Resource
-
-	valResources map[string][]*dockertest.Resource
-}
-
-type AddressResponse struct {
-	Name     string `json:"name"`
-	Type     string `json:"type"`
-	Address  string `json:"address"`
-	Mnemonic string `json:"mnemonic"`
-}
 
 func TestIntegrationTestSuite(t *testing.T) {
 	suite.Run(t, new(IntegrationTestSuite))
@@ -680,213 +599,6 @@ func (s *IntegrationTestSuite) runIBCRelayer() {
 	// create the client, connection and channel between the two Gaia chains
 	s.createConnection()
 	s.createChannel()
-}
-
-func (s *IntegrationTestSuite) writeGovCommunitySpendProposal(c *chain, amount sdk.Coin, recipient string) {
-	template := `
-	{
-		"messages":[
-		  {
-			"@type": "/cosmos.distribution.v1beta1.MsgCommunityPoolSpend",
-			"authority": "%s",
-			"recipient": "%s",
-			"amount": [{
-				"denom": "%s",
-				"amount": "%s"
-			}]
-		  }
-		],
-		"deposit": "100uatom",
-		"proposer": "Proposing validator address",
-		"metadata": "Community Pool Spend",
-		"title": "Fund Team!",
-		"summary": "summary",
-		"expedited": false
-	}
-	`
-	propMsgBody := fmt.Sprintf(template, govModuleAddress, recipient, amount.Denom, amount.Amount.String())
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalCommunitySpendFilename), []byte(propMsgBody))
-	s.Require().NoError(err)
-}
-
-func (s *IntegrationTestSuite) writeSoftwareUpgradeProposal(c *chain, height int64, name string) {
-	body := `{
-		"messages": [
-		 {
-		  "@type": "/cosmos.upgrade.v1beta1.MsgSoftwareUpgrade",
-		  "authority": "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn",
-		  "plan": {
-		   "name": "%s",
-		   "height": "%d",
-		   "info": "test",
-		   "upgraded_client_state": null
-		  }
-		 }
-		],
-		"metadata": "ipfs://CID",
-		"deposit": "100uatom",
-		"title": "title",
-		"summary": "test"
-	   }`
-
-	propMsgBody := fmt.Sprintf(body, name, height)
-
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalSoftwareUpgrade), []byte(propMsgBody))
-	s.Require().NoError(err)
-}
-
-func (s *IntegrationTestSuite) writeCancelSoftwareUpgradeProposal(c *chain) {
-	template := `{
-		"messages": [
-		 {
-		  "@type": "/cosmos.upgrade.v1beta1.MsgCancelUpgrade",
-		  "authority": "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn"
-		 }
-		],
-		"metadata": "ipfs://CID",
-		"deposit": "100uatom",
-		"title": "title",
-		"summary": "test"
-	   }`
-
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalCancelSoftwareUpgrade), []byte(template))
-	s.Require().NoError(err)
-}
-
-func (s *IntegrationTestSuite) writeLiquidStakingParamsUpdateProposal(c *chain, oldParams stakingtypes.Params) {
-	template := `
-	{
-		"messages": [
-		 {
-		  "@type": "/cosmos.staking.v1beta1.MsgUpdateParams",
-		  "authority": "%s",
-		  "params": {
-		   "unbonding_time": "%s",
-		   "max_validators": %d,
-		   "max_entries": %d,
-		   "historical_entries": %d,
-		   "bond_denom": "%s",
-		   "min_commission_rate": "%s",
-		   "validator_bond_factor": "%s",
-		   "global_liquid_staking_cap": "%s",
-		   "validator_liquid_staking_cap": "%s"
-		  }
-		 }
-		],
-		"metadata": "ipfs://CID",
-		"deposit": "100uatom",
-		"title": "Update LSM Params",
-		"summary": "e2e-test updating LSM staking params",
-		"expedited": false
-	   }`
-	propMsgBody := fmt.Sprintf(template,
-		govAuthority,
-		oldParams.UnbondingTime,
-		oldParams.MaxValidators,
-		oldParams.MaxEntries,
-		oldParams.HistoricalEntries,
-		oldParams.BondDenom,
-		oldParams.MinCommissionRate,
-		math.LegacyNewDec(250),           // validator bond factor
-		math.LegacyNewDecWithPrec(25, 2), // 25 global_liquid_staking_cap
-		math.LegacyNewDecWithPrec(50, 2), // 50 validator_liquid_staking_cap
-	)
-
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalLSMParamUpdateFilename), []byte(propMsgBody))
-	s.Require().NoError(err)
-}
-
-// writeGovParamChangeProposalBlocksPerEpoch writes a governance proposal JSON file to change the `BlocksPerEpoch`
-// parameter to the provided `blocksPerEpoch`
-func (s *IntegrationTestSuite) writeGovParamChangeProposalBlocksPerEpoch(c *chain, paramsJSON string) {
-	template := `
-	{
-		"messages":[
-		  {
-			"@type": "/interchain_security.ccv.provider.v1.MsgUpdateParams",
-   			"authority": "%s",
-			"params": %s
-		  }
-		],
-		"deposit": "100uatom",
-		"proposer": "sample proposer",
-		"metadata": "sample metadata",
-		"title": "blocks per epoch title",
-		"summary": "blocks per epoch summary",
-		"expedited": false
-	}`
-
-	propMsgBody := fmt.Sprintf(template,
-		govAuthority,
-		paramsJSON,
-	)
-
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalBlocksPerEpochFilename), []byte(propMsgBody))
-	s.Require().NoError(err)
-}
-
-// writeFailingExpeditedProposal writes a governance proposal JSON file.
-// The proposal fails because only SoftwareUpgrade and CancelSoftwareUpgrade can be expedited.
-func (s *IntegrationTestSuite) writeFailingExpeditedProposal(c *chain, blocksPerEpoch int64) {
-	template := `
-	{
-		"messages":[
-		  {
-			"@type": "/cosmos.gov.v1.MsgExecLegacyContent",
-			"authority": "%s",
-			"content": {
-				"@type": "/cosmos.params.v1beta1.ParameterChangeProposal",
-				"title": "BlocksPerEpoch",
-				"description": "change blocks per epoch",
-				"changes": [{
-				  "subspace": "provider",
-				  "key": "BlocksPerEpoch",
-				  "value": "\"%d\""
-				}]
-			}
-		  }
-		],
-		"deposit": "100uatom",
-		"proposer": "sample proposer",
-		"metadata": "sample metadata",
-		"title": "blocks per epoch title",
-		"summary": "blocks per epoch summary",
-		"expedited": true
-	}`
-
-	propMsgBody := fmt.Sprintf(template,
-		govAuthority,
-		blocksPerEpoch,
-	)
-
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalFailExpedited), []byte(propMsgBody))
-	s.Require().NoError(err)
-}
-
-// MsgSoftwareUpgrade can be expedited and it can only be submitted using "tx gov submit-proposal" command.
-func (s *IntegrationTestSuite) writeExpeditedSoftwareUpgradeProp(c *chain) {
-	body := `{
- "messages": [
-  {
-   "@type": "/cosmos.upgrade.v1beta1.MsgSoftwareUpgrade",
-   "authority": "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn",
-   "plan": {
-    "name": "test-expedited-upgrade",
-    "height": "123456789",
-    "info": "test",
-    "upgraded_client_state": null
-   }
-  }
- ],
- "metadata": "ipfs://CID",
- "deposit": "100uatom",
- "title": "title",
- "summary": "test",
- "expedited": true
-}`
-
-	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalExpeditedSoftwareUpgrade), []byte(body))
-	s.Require().NoError(err)
 }
 
 func configFile(filename string) string {
