+++ tests/e2e/test.go
@@ -1,0 +1,1 @@
+package e2e
